models: run on Databricks' samples data, with provenance to the durable record
All five models now read the free `samples` catalog when there is a workspace
and fall back deterministically when there is not. Built as five parallel
tracks against models/_data, which landed first for the same reason shared/
did — every track needed it frozen before it could start.

- forecasting: trains on hourly NYC taxi volumes; drops rows whose target is
  NULL, because the real table has them and a lag window built on one is
  worthless. Results carry the forecast hour as a real timestamp.
- gurobi_scheduling: coverage requirements derived from the real demand curve.
  840 variables / 602 constraints, unchanged — demand only moves right-hand
  sides. A test sweeps 1 to 500,000 trips/hour and asserts the model stays
  under the licence cap AND stays solvable, so the clipping produces
  feasibility rather than merely a smaller number.
- mcmc: fare ~ distance regression over real trips; the analytic toy problem
  stays as the correctness fixture, since a known posterior is how the suite
  checks the sampler lands somewhere right rather than merely somewhere.
- scenario: baseline from observed demand — mean hourly trips for demand, p90
  for capacity (what a fleet is sized for, not the mean hour), mean fare for
  unit cost.
- streaming_results: rolling-origin backtest over the real series, and a new
  edge case: too short a series used to emit zero chunks and therefore no
  final=True at all, leaving a client unable to tell "complete, no rows" from
  "still coming". It now emits one empty final chunk.

Three defects in models/_data, all found by the tracks and all mine:

- hour_ts was documented as epoch ms but `date_trunc` returns a Spark
  TimestampType, which comes back as a datetime — while the fallback returned
  an int. forecasting's `int(row["hour_ts"])` would have worked in every test
  and raised on the first real run. Cast in SQL now, and normalised in Python
  as belt and braces.
- floats() called float(None), which a real NULL aggregate triggers.
- describe() omitted data_fallback_reason on success, giving one table two row
  schemas; all four tracks wrote the same workaround, which was the signal.

Also: uc_ddl declared data_rows as INT for one model and BIGINT for four —
one logical field, two types, which is the drift concurrent editing of a
shared file produces. And CREATE TABLE IF NOT EXISTS will not add a column to
an existing table, so uc_ddl/README.md now spells out the ALTER TABLE, and why
it bites the Spark writer but not delta-rs.

342 tests, ruff clean, ty clean, 30 deploy contract tests green.

+++ uc_ddl/002_model_results.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS main.dbx_leaning.results_gurobi_scheduling (
     demand               INT,
     data_source          STRING,
     data_synthetic       BOOLEAN,
-    data_rows            INT,
+    data_rows            BIGINT,
     data_fallback_reason STRING
 )
 USING DELTA
